requirements.txt
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
openai>=1.12.0
pgvector>=0.2.4
python-dotenv>=1.0.0
pydantic>=2.5.0
